+++ contracts/mpc/PodUser.sol
@@ -4,13 +4,10 @@ pragma solidity ^0.8.19;
 import "../IInbox.sol";
 import "../InboxUser.sol";
 
-/**
- * @title PodUser
- * @notice Base for POD contracts that talk to the inbox: configure the COTI MPC executor
- *         and send two-way messages to MpcExecutor on the COTI side.
- */
+/// @title PodUser
+/// @notice POD base: COTI chain ID, MPC executor address, and configuration hook.
 abstract contract PodUser is InboxUser {
-    event ErrorRemoteCall(bytes32 requestId, uint code, string message);
+    event ErrorRemoteCall(bytes32 requestId, uint256 code, string message);
 
     address internal mpcExecutorAddress = 0x0000000000000000000000000000000000000000;
     uint256 internal cotiChainId = 2632500;
